refactor: Main 페이지 기능에 따른 util 함수 분리
